+++ yandex/data_source_yandex_logging_group_test.go
@@ -2,6 +2,7 @@ package yandex
 
 import (
 	"fmt"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"math/rand"
 	"testing"
 	"time"
@@ -12,7 +13,12 @@ import (
 	"github.com/yandex-cloud/go-genproto/yandex/cloud/logging/v1"
 )
 
-const yandexLoggingGroupDataSource = "data.yandex_logging_group.test-logging-group"
+const (
+	yandexLoggingGroupDataSource = "data.yandex_logging_group.test-logging-group"
+	dataStreamName               = "logging-yds"
+	ydbResource                  = "logging-ydb"
+	topicResource                = "logging-topic"
+)
 
 func TestAccDataSourceYandexLoggingGroup_byID(t *testing.T) {
 	var group logging.LogGroup
@@ -75,6 +81,7 @@ func TestAccDataSourceYandexLoggingGroup_full(t *testing.T) {
 	params := testYandexLoggingGroupParameters{
 		name:            acctest.RandomWithPrefix("tf-yandex-logging-group"),
 		desc:            acctest.RandomWithPrefix("tf-yandex-logging-group-desc"),
+		dataStream:      dataStreamName,
 		labelKey:        acctest.RandomWithPrefix("tf-yandex-logging-group-label"),
 		labelValue:      acctest.RandomWithPrefix("tf-yandex-logging-group-label-value"),
 		retentionPeriod: time.Hour + time.Duration(rand.Uint32())*time.Nanosecond,
@@ -87,17 +94,22 @@ func TestAccDataSourceYandexLoggingGroup_full(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				Config: testYandexLoggingGroupDataSource(params),
-				Check: resource.ComposeTestCheckFunc(
-					testYandexLoggingGroupExists(yandexLoggingGroupDataSource, &group),
-					resource.TestCheckResourceAttr(yandexLoggingGroupDataSource, "name", params.name),
-					resource.TestCheckResourceAttr(yandexLoggingGroupDataSource, "description", params.desc),
-					resource.TestCheckResourceAttr(yandexLoggingGroupDataSource, "retention_period", params.retentionPeriod.String()),
-					resource.TestCheckResourceAttrSet(yandexLoggingGroupDataSource, "folder_id"),
-					resource.TestCheckResourceAttrSet(yandexLoggingGroupDataSource, "cloud_id"),
-					resource.TestCheckResourceAttrSet(yandexLoggingGroupDataSource, "created_at"),
-					testYandexLoggingGroupContainsLabel(&group, params.labelKey, params.labelValue),
-					testAccCheckCreatedAtAttr(yandexLoggingGroupDataSource),
-				),
+				Check: func(s *terraform.State) error {
+					databasePath := s.RootModule().Resources["yandex_ydb_database_serverless."+ydbResource].Primary.Attributes["database_path"]
+					dataStreamFullName := databasePath + "/" + params.dataStream
+					return resource.ComposeTestCheckFunc(
+						testYandexLoggingGroupExists(yandexLoggingGroupDataSource, &group),
+						resource.TestCheckResourceAttr(yandexLoggingGroupDataSource, "name", params.name),
+						resource.TestCheckResourceAttr(yandexLoggingGroupDataSource, "description", params.desc),
+						resource.TestCheckResourceAttr(yandexLoggingGroupDataSource, "data_stream", dataStreamFullName),
+						resource.TestCheckResourceAttr(yandexLoggingGroupDataSource, "retention_period", params.retentionPeriod.String()),
+						resource.TestCheckResourceAttrSet(yandexLoggingGroupDataSource, "folder_id"),
+						resource.TestCheckResourceAttrSet(yandexLoggingGroupDataSource, "cloud_id"),
+						resource.TestCheckResourceAttrSet(yandexLoggingGroupDataSource, "created_at"),
+						testYandexLoggingGroupContainsLabel(&group, params.labelKey, params.labelValue),
+						testAccCheckCreatedAtAttr(yandexLoggingGroupDataSource),
+					)(s)
+				},
 			},
 		},
 	})
@@ -134,6 +146,16 @@ data "yandex_logging_group" "test-logging-group" {
   group_id = "${yandex_logging_group.test-logging-group.id}"
 }
 
+resource "yandex_ydb_database_serverless" "%s" {
+	name = "%s"
+	location_id = "ru-central1"
+}
+
+resource "yandex_ydb_topic" "%s" {
+	name = "%s"
+	database_endpoint = "${yandex_ydb_database_serverless.%s.ydb_full_endpoint}"
+}
+
 resource "yandex_logging_group" "test-logging-group" {
   name        = "%s"
   description = "%s"
@@ -142,12 +164,20 @@ resource "yandex_logging_group" "test-logging-group" {
     empty-label = ""
   }
   retention_period = "%s"
+  data_stream = "${yandex_ydb_database_serverless.%s.database_path}/%s"
 }
 `,
+		ydbResource,
+		ydbResource,
+		topicResource,
+		dataStreamName,
+		ydbResource,
 		params.name,
 		params.desc,
 		params.labelKey,
 		params.labelValue,
 		params.retentionPeriod,
+		ydbResource,
+		dataStreamName,
 	)
 }